+++ src/view/page/ModInputElement/ModInputElement.tsx
@@ -1,17 +1,12 @@
-/** @jsxImportSource @emotion/react */
-import { css } from '@emotion/react';
-
-import { Modal, InputNumber } from 'antd';
-import { TTheme } from '../../../types';
+import { Modal } from 'antd';
 import { useModInputViewModel } from '../../../viewModels';
-import Paragraph from 'antd/es/typography/Paragraph';
 import { ModalFooter } from '../../elements';
 import { Buttons } from './Buttons';
+import { ModalBody } from './ModelBody';
 
 export const ModInputElement = () => {
     const { theme, toggleShowInput, translations, showInput, currentValue, updateCurrentValue, onConfirm, onRemove } =
         useModInputViewModel();
-    const themedStyles = styles(theme);
 
     const onChange = (value: number | null) => {
         value && updateCurrentValue(value);
@@ -40,32 +35,8 @@ export const ModInputElement = () => {
                     />
                 }
             >
-                <div css={themedStyles.inputWrapper}>
-                    <InputNumber onChange={onChange} value={currentValue} css={themedStyles.inputWrapper} />
-                </div>
-                <Paragraph type="warning" css={themedStyles.warning}>
-                    {translations.MOD_TIP}
-                </Paragraph>
+                <ModalBody theme={theme} translations={translations} currentValue={currentValue} onChange={onChange} />
             </Modal>
         </>
     );
 };
-
-const styles = (theme: TTheme) => ({
-    warning: css`
-        font-size: ${0.8 * theme.fontSize}px;
-        font-style: italic;
-        text-align: center;
-    `,
-    inputWrapper: css`
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin: ${3 * theme.baseSpace}px 0;
-    `,
-    input: css`
-        .ant-input-number-input {
-            text-align: center;
-        }
-    `,
-});
